Keep AG-UI decoding helpers easier to reuse
The AG-UI decoder mixed newline normalization, custom chunk shaping, SSE frame splitting, finish-reason mapping, and tool-result parsing directly into the main module, so this extracts the pure decoding helpers into a sibling module without changing emitted chat events.

Constraint: Refactor must preserve AG-UI decoding semantics and emitted chat stream events exactly
Constraint: veryfront-code pre-commit verify:quick currently fails on pre-existing CLI boundary violations in cli/commands/extension/*.ts
Rejected: Move reasoning fallback state helpers too | would widen the change into stateful decoder flow instead of keeping this pass purely functional
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep pure AG-UI decoding utilities in the helper module unless a future change needs stateful logic tied to the main decoder
Tested: deno test --no-check --allow-all src/chat/ag-ui.test.ts; deno fmt --check src/chat/ag-ui.ts src/chat/ag-ui-helpers.ts; deno lint src/chat/ag-ui.ts src/chat/ag-ui-helpers.ts src/chat/ag-ui.test.ts; git diff --check
Not-tested: Full veryfront-code repo suite
