+++ src/chat/ag-ui.ts
@@ -1,5 +1,15 @@
 import { mergeToolInputDelta, parseToolInputObject } from "../agent/data-stream.ts";
-import type { ChatFinishReason, ChatStreamEvent } from "./protocol.ts";
+import {
+  formatToolErrorText,
+  isCommentOnlySseFrame,
+  isRecord,
+  mapFinishReason,
+  normalizeNewlines,
+  parseSerializedToolResult,
+  splitSseFrames,
+  toRenderableCustomChunk,
+} from "./ag-ui-helpers.ts";
+import type { ChatStreamEvent } from "./protocol.ts";
 import { z } from "zod";
 
 type JsonPatchOperation = {
@@ -8,11 +18,6 @@ type JsonPatchOperation = {
   from?: string;
   value?: unknown;
 };
-
-type ParsedRenderableCustomChunk = Extract<
-  ChatStreamEvent,
-  { type: "source-url" | "source-document" | "file" }
->;
 
 type ToolCallState = {
   toolName: string;
@@ -291,125 +296,6 @@ export type AgUiSnapshotMessage = z.infer<typeof AgUiSnapshotMessageSchema>;
 export type AgUiWireEventName = z.infer<typeof AgUiWireEventNameSchema>;
 export type AgUiWireEvent = z.infer<typeof AgUiWireEventSchema>;
 
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
-function normalizeNewlines(value: string): string {
-  return value.replace(/\r\n/g, "\n").replace(/\r/g, "\n");
-}
-
-function toRenderableCustomChunk(value: unknown): ParsedRenderableCustomChunk | null {
-  if (!isRecord(value) || typeof value.type !== "string") {
-    return null;
-  }
-
-  if (value.type === "source-url" && typeof value.url === "string") {
-    return {
-      type: "source-url",
-      sourceId: typeof value.sourceId === "string" && value.sourceId.length > 0
-        ? value.sourceId
-        : value.url,
-      url: value.url,
-      ...(typeof value.title === "string" ? { title: value.title } : {}),
-    };
-  }
-
-  if (
-    value.type === "source-document" &&
-    typeof value.sourceId === "string" &&
-    typeof value.mediaType === "string" &&
-    typeof value.title === "string"
-  ) {
-    return {
-      type: "source-document",
-      sourceId: value.sourceId,
-      mediaType: value.mediaType,
-      title: value.title,
-      ...(typeof value.filename === "string" ? { filename: value.filename } : {}),
-    };
-  }
-
-  if (
-    value.type === "file" && typeof value.url === "string" && typeof value.mediaType === "string"
-  ) {
-    return {
-      type: "file",
-      url: value.url,
-      mediaType: value.mediaType,
-      ...(typeof value.filename === "string" ? { filename: value.filename } : {}),
-    };
-  }
-
-  return null;
-}
-
-function parseSerializedToolResult(value: unknown): unknown {
-  if (typeof value !== "string") {
-    return value;
-  }
-
-  const trimmed = value.trim();
-  if (
-    !trimmed.startsWith("{") &&
-    !trimmed.startsWith("[") &&
-    trimmed !== "null" &&
-    trimmed !== "true" &&
-    trimmed !== "false" &&
-    !/^[-]?\d+(\.\d+)?$/.test(trimmed)
-  ) {
-    return value;
-  }
-
-  try {
-    return JSON.parse(trimmed);
-  } catch {
-    return value;
-  }
-}
-
-function formatToolErrorText(result: unknown): string {
-  if (typeof result === "string" && result.length > 0) {
-    return result;
-  }
-
-  if (isRecord(result)) {
-    if (typeof result.error === "string" && result.error.length > 0) {
-      return result.error;
-    }
-
-    if (typeof result.message === "string" && result.message.length > 0) {
-      return result.message;
-    }
-  }
-
-  return JSON.stringify(result ?? { error: "Tool execution failed" });
-}
-
-function mapFinishReason(reason: string | undefined): ChatFinishReason | undefined {
-  if (!reason) return undefined;
-
-  switch (reason.trim().toLowerCase()) {
-    case "stop":
-    case "end_turn":
-    case "stop_sequence":
-      return "stop";
-    case "length":
-    case "max_tokens":
-      return "length";
-    case "tool_calls":
-    case "tool_use":
-      return "tool-calls";
-    case "content_filter":
-    case "content-filter":
-      return "content-filter";
-    case "error":
-      return "error";
-    default:
-      return "other";
-  }
-}
-
 function getReasoningPartId(
   state: AgUiChatEventDecoderState,
   payload: { id?: string; messageId?: string },
@@ -437,20 +323,6 @@ function getReasoningPartId(
     state.activeFallbackReasoningPartId = fallbackId;
   }
   return fallbackId;
-}
-
-function splitSseFrames(value: string): { frames: string[]; remainder: string } {
-  const blocks = value.split("\n\n");
-  return {
-    frames: blocks.slice(0, -1),
-    remainder: blocks.at(-1) ?? "",
-  };
-}
-
-function isCommentOnlySseFrame(raw: string): boolean {
-  return raw
-    .split("\n")
-    .every((line) => line.trim().length === 0 || line.trimStart().startsWith(":"));
 }
 
 function parseAgUiWireEvent(
